splitted background.js file to make code more manageable
